refactor(cli): extract query helpers to shared module
Move 14 duplicated helper functions from list.ts and index.ts to
apps/cli/src/query-helpers.ts:
- Parsing: parseCsvList, parsePrList, parseTypes, parseAuthorFilters
- Resolution: resolveBotPatterns, resolveSinceFilter, resolveAuthorFilters,
  resolveRepoFilter, resolveReposToSync
- Cache utilities: isFullRepo, listCachedRepos, getSyncMetaMap,
  hasRepoCache, isStale, ensureRepoCache, ensureFreshRepos
- Global options: applyGlobalOptions, DEFAULT_STALE_THRESHOLD

Eliminates ~400 lines of duplication. Complexity warnings remain
(addressed in next branch).
